+++ cmd/centre/dhcp_darwin.go
@@ -1,0 +1,7 @@
+package main
+
+import "log"
+
+func dhcp() {
+	log.Fatal("No DHCP on this kernel yet")
+}
